Changes:
Contrary to what we thought, Coq supports mutual recursion quite
well. Changed syntax and type checking definitions to utilize mutual
recursion. Simplified (a little) most of the proofs.

The subtype definition was wrong. It should have been Type when it was
Prop. Because of that, automatic naming of local context variables
changed for H* to X* for subtypes. Since And/Or is Prop -> Prop -> Prop,
WF_Var was wrong and had to change to Prop. WF_Var is used everywhere in
the proofs, and most proofs used automatic naming. Many small changes
were needed in many of the proofs.